It works!
Update 5/15/20:
It works! It is completely functional, but doesn't have the difficulty setting where it is impossible to win yet. 

Added the functionality to go first or second. 

Added the easy difficulty, where the bot chooses randomly based off of empty squares.
Added emptySquares() to determine what those squares were.

Added the medium difficulty, which prioritized centers and corners.

Added the hard difficulty, which was the same as the medium difficulty, but also had a functionality to always choose a guaranteed win if it was available, or to always prevent a guaranteed loss.

TODO: Add impossible difficulty, where the bot always wins or ties.